refactor(web): move components to grouping folder
